Move pkg/helmchart to pkg/input/helmchart
