+++ cmd/root.go
@@ -13,7 +13,7 @@ import (
 	"golang.org/x/oauth2"
 
 	"github.com/giantswarm/backstage-catalog-importer/pkg/export"
-	"github.com/giantswarm/backstage-catalog-importer/pkg/helmchart"
+	"github.com/giantswarm/backstage-catalog-importer/pkg/input/helmchart"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/teams"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/legacy"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/repositories"
